% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_date.R
\name{as_date}
\alias{as_date}
\title{as_date}
\usage{
as_date(x, format = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{Character or factor vector of dates.}

\item{format}{Date format string.}
}
\description{
Faster alternative to as.Date. Converts character vectors to factors
and converts unique values to dates.
}
\examples{
as_date("2010-01-01")

}
