% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{clipboard}
\alias{clipboard}
\title{clipboard}
\usage{
clipboard(x = NULL, alloc = 2048)
}
\arguments{
\item{x}{A data.frame.}

\item{alloc}{Memory to allocate for copied data.frame. Should be a power of
2.}
}
\value{
A data.frame
}
\description{
clipboard
}
\examples{
clipboard(mtcars)
print(clipboard())
}
