% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_lines.R
\name{count_lines}
\alias{count_lines}
\title{count_lines}
\usage{
count_lines(file, n = 10000, verbose = FALSE)
}
\arguments{
\item{file}{path to text file}

\item{n}{chunk size to read in at a time}

\item{verbose}{logical}
}
\description{
counts the number of lines in a text file
}
