% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_stack.R
\name{dir_stack}
\alias{dir_stack}
\title{dir_stack}
\usage{
dir_stack(path = ".", pattern = "\\\\.csv$", file_paths = NULL,
  ind = NULL, recursive = FALSE, func = NULL, read_func = function(x)
  read.csv(x, stringsAsFactors = FALSE), verbose = FALSE)
}
\arguments{
\item{path}{Path to a directory.}

\item{pattern}{Regular expression to match on file names.}

\item{file_paths}{A vector of files paths. If supplied, /code{path},
/code{pattern} and \code{recursive} are ignored.}

\item{ind}{Logical or numeric indices used to subset file_paths.}

\item{recursive}{passed to \code{list.files()}}

\item{func}{A function for processing object read in by \code{read_func()}}

\item{read_func}{A function for reading in files specified in \code{file_paths}.}

\item{verbose}{\code{TRUE} or \code{FALSE}. Default is \code{FALSE}.}
}
\value{
A data.frame.
}
\description{
dir_stack
}
