% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downfactor.R
\name{downfactor}
\alias{downfactor}
\title{downfactor}
\usage{
downfactor(x, y = rep(1, length(x)), n = 5)
}
\arguments{
\item{x}{A character or factor vector}

\item{y}{A numeric vector}

\item{n}{Number of factors to retain}
}
\description{
downfactor
}
