% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_date.R
\name{fast_date}
\alias{fast_date}
\title{fast_date}
\usage{
fast_date(x, format = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{Character or factor vector of dates.}

\item{format}{Date format string.}
}
\description{
Faster alternative to as.Date. Converts character vectors to factors
and converts unique values to dates.
}
\examples{
fast_date("2010-01-01")

}
