% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_headers.R
\name{get_headers}
\alias{get_headers}
\title{get_headers}
\usage{
get_headers(data, header_row, cols, fill = TRUE, make.names = TRUE)
}
\arguments{
\item{data}{A data.frame.}

\item{header_row}{A scalar numeric (row containing headers).}

\item{cols}{A numeric vector (column indices).}
}
\description{
extracts header names from unparsed data.frame (usually from Excel)
}
