% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movav.R
\name{movav}
\alias{movav}
\title{Moving average of a vector}
\usage{
movav(x, n = 52, na.rm = TRUE, na.start = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{Number of values to average.}

\item{na.rm}{A logical. Whether or not to treat NAs as 0.}

\item{na.start}{Whether starting elements without enough values to average
should calculate to NA. Default is \code{FALSE}.}
}
\description{
Calculates the moving average of a vector.
}
\examples{
movav(runif(100))
}
