% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_f.R
\name{na_f}
\alias{na_f}
\title{na_f}
\usage{
na_f(x, f)
}
\arguments{
\item{x}{vector}

\item{f}{function}
}
\description{
run common functions but force all-NA inputs to return NA output
}
\examples{
lqpm:::na(NA, sum)
lqpm:::na(1, sum)
lqpm:::na(as.Date(NA), min)
lqpm:::na(as.Date("2010-01-01"), min)
}
