% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel2.R
\name{read_excel2}
\alias{read_excel2}
\title{read_excel2}
\usage{
read_excel2(path, sheet = NULL, data_row = 2,
  cat_header_row = data_row - 1, val_header_row = data_row - 1,
  first_val_col = NULL, gather = FALSE)
}
\description{
read_excel2
}
\examples{
file = "inst/extdata/read_excel2.xlsx"
sheet = "format1"
data_row = 4
read_excel2(file, sheet, data_row, val_header_row = 2)
file = "inst/extdata/read_excel2.xlsx"
sheet = "format2"
data_row = 5
read_excel2(file, sheet, data_row, val_header_row = 3, gather = TRUE)
}
