% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_left_join.R
\name{safe_left_join}
\alias{safe_left_join}
\title{safe_left_join}
\usage{
safe_left_join(x, y, by = NULL, verbose = TRUE)
}
\arguments{
\item{x}{table to join}

\item{y}{table to join}

\item{by}{a character vector of column names to join by.}

\item{verbose}{Default is TRUE.}
}
\description{
left join that fails if a row in x is either duplicated or
  unmatched.
}
