% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_file.R
\name{scan_file}
\alias{scan_file}
\title{scan_file}
\usage{
scan_file(file, n = 10000, read_func = function(file, nrows, header) {    
  read.csv(file, nrows = nrows, header = header) }, header = FALSE,
  agg_func = econ::side_summary, verbose = FALSE)
}
\arguments{
\item{file}{path to a text file}

\item{n}{chunk size}

\item{read_func}{function to use for reading file}

\item{header}{logical indicating whether file has a header}

\item{agg_func}{aggregation function for each chunk}

\item{verbose}{logical}
}
\description{
scan_file
}
