% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsummary.R
\name{xsummary}
\alias{xsummary}
\title{xsummary}
\usage{
xsummary(..., sample = 2)
}
\arguments{
\item{...}{Vectors for intersection summary}

\item{sample}{Number of samples to show from each group.}
}
\description{
intersection summary
}
