% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_part.R
\name{year_part}
\alias{year_part}
\title{year_part}
\usage{
year_part(x)
}
\arguments{
\item{x}{A date vector.}
}
\description{
convert a date vector to labelled year parts
}
\examples{
year_part(seq(as.Date("2001-01-01"), by = "month", length.out = 16))
}
