% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_elements.R
\name{count_elements}
\alias{count_elements}
\title{count_elements}
\usage{
count_elements(x, sep = ",", quote = NULL)
}
\arguments{
\item{x}{A character vector. Usually, the output from \code{readLines()}.}

\item{sep}{A separator used in \code{x}.}
}
\description{
count_elements
}
\examples{
text = c("a,b", "b,d", 'a,",,d"', "a, b, c, d")
count_elements(text)
}
