% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_elements.R
\name{count_elements}
\alias{count_elements}
\title{count_elements}
\usage{
count_elements(text, sep = ",", quote = "\\"")
}
\arguments{
\item{sep}{A separator used in \code{x}.}

\item{x}{A character vector. Usually, the output from \code{readLines()}.}
}
\description{
count_elements
}
\examples{
text = c("a,b,c", 'a,",,b",c')
count_elements(text)
count_elements(text, ",", '')
}
