% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_profile.R
\name{na_profile}
\alias{na_profile}
\title{na_profile}
\usage{
na_profile(data, all_cols = FALSE)
}
\description{
na_profile
}
\examples{
na_profile(airquality)
na_profile(airquality, all_cols = TRUE)
na_profile(mtcars)
na_profile(mtcars, all_cols = TRUE)
}
