% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_summary.R
\name{side_summary}
\alias{side_summary}
\title{side_summary}
\usage{
side_summary(data, ..., unique = FALSE, sample = FALSE)
}
\arguments{
\item{data}{A data.frame.}

\item{...}{Extra inputs.}
}
\description{
side_summary
}
\examples{
side_summary(mtcars)
side_summary(mtcars, unique = T)
side_summary(mtcars, unique = T, sample = T)
}
